RAG_local/hybrid/requirements.txt
openai>=1.0.0
pinecone-client>=2.2.1
rank_bm25>=0.2.2
python-dotenv>=0.19.0